// src/app/models/lastSelection.model.ts
export class LastSelection {
      accountId: number;
      associationId: number;
      creditCardAccId: number;
      paymentFor: number;
      paymenthMethodId: number;
      project: number;
      projectCategory: number;
      receiptTypeId: number;
      receiptFor: string;
      receiptTemplate: number;
      selectedOrg: number;
      selected_receiptCreditOrDebit: boolean;
      selected_receiptIsForDonation: boolean;
}